feat(ui): move expired image URL handling into hook
